chore: cleaning up code
- Moving all types to the types file
- Moving error functions to the errors file
- Moving validation functions to a validation file
- Moving timeout function to a timeout file
