+++ src/errors.ts
@@ -1,3 +1,5 @@
+import { AWSError } from './types';
+
 class SQSError extends Error {
   code: string;
   statusCode: number;
@@ -20,4 +22,28 @@ class TimeoutError extends Error {
   }
 }
 
-export { SQSError, TimeoutError };
+function isConnectionError(err: Error): boolean {
+  if (err instanceof SQSError) {
+    return (
+      err.statusCode === 403 ||
+      err.code === 'CredentialsError' ||
+      err.code === 'UnknownEndpoint' ||
+      err.code === 'AWS.SimpleQueueService.NonExistentQueue'
+    );
+  }
+  return false;
+}
+
+function toSQSError(err: AWSError, message: string): SQSError {
+  const sqsError = new SQSError(message);
+  sqsError.code = err.name;
+  sqsError.statusCode = err.$metadata?.httpStatusCode;
+  sqsError.retryable = err.$retryable?.throttling;
+  sqsError.service = err.$service;
+  sqsError.fault = err.$fault;
+  sqsError.time = new Date();
+
+  return sqsError;
+}
+
+export { SQSError, TimeoutError, isConnectionError, toSQSError };
