+++ src/timeout.ts
@@ -1,0 +1,14 @@
+import { TimeoutResponse } from './types';
+import { TimeoutError } from './errors';
+
+function createTimeout(duration: number): TimeoutResponse[] {
+  let timeout;
+  const pending = new Promise((_, reject) => {
+    timeout = setTimeout((): void => {
+      reject(new TimeoutError());
+    }, duration);
+  });
+  return [timeout, pending];
+}
+
+export { createTimeout };
